Style cart for mobile
