+++ src/routes/(app)/shop/cart/CartItem/CartItemRemoveButton.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+  import { enhance } from "$app/forms";
+
+  export let itemId: string;
+</script>
+
+<form method="POST" action="?/removeItem" use:enhance>
+  <input type="hidden" name="id" value={itemId} />
+  <button type="submit" class="btn btn-ghost">
+    <span class="i-mdi-trash-can text-xl" />
+  </button>
+</form>
