+++ src/routes/(app)/shop/cart/payments/PurchaseSection.svelte
@@ -1,0 +1,53 @@
+<script lang="ts">
+  import { PUBLIC_STRIPE_KEY } from "$env/static/public";
+  import type StripeJS from "@stripe/stripe-js";
+  import { loadStripe } from "@stripe/stripe-js";
+  import { onMount } from "svelte";
+  import type { SuperValidated } from "sveltekit-superforms";
+  import { superForm } from "sveltekit-superforms/client";
+  import type { PurchaseForm } from "../+page.server";
+  import SveltePaymentElement from "./SveltePaymentElement.svelte";
+  import Price from "$lib/components/Price.svelte";
+
+  export let totalPrice: number;
+
+  let stripe: StripeJS.Stripe | null = null;
+  onMount(async () => {
+    stripe = await loadStripe(PUBLIC_STRIPE_KEY);
+  });
+
+  const idempotencyKey = crypto.randomUUID();
+
+  export let purchaseForm: SuperValidated<PurchaseForm>;
+  const { enhance, message, submitting } = superForm(purchaseForm);
+</script>
+
+{#if $message?.["clientSecret"] !== undefined}
+  <tr>
+    <td colspan="3" class="w-full max-w-xl">
+      <SveltePaymentElement
+        {stripe}
+        clientSecret={$message["clientSecret"]}
+        price={totalPrice}
+      />
+    </td>
+  </tr>
+{:else}
+  <tr>
+    <td>
+      <form method="POST" action="?/purchase" use:enhance>
+        <input type="hidden" name="idempotencyKey" value={idempotencyKey} />
+        <button type="submit" class="btn btn-primary" disabled={$submitting}>
+          {#if $submitting}
+            Processar...
+          {:else}
+            {totalPrice === 0 ? "Skaffa" : "Betala"}
+          {/if}
+        </button>
+      </form>
+    </td>
+    <td colspan="2" class="text-right">
+      <Price price={totalPrice} />
+    </td>
+  </tr>
+{/if}
