+++ src/routes/(app)/shop/tickets/[slug]/+page@(app).svelte
@@ -2,6 +2,7 @@
   import { enhance } from "$app/forms";
   import dayjs from "dayjs";
   import BuyButton from "$lib/components/BuyButton.svelte";
+  import Price from "$lib/components/Price.svelte";
 
   export let data;
   $: ticket = data.ticket;
@@ -21,9 +22,7 @@
   <main class="m-4 flex flex-col gap-4">
     <h1 class="text-xl font-bold">
       {ticket.title}
-      <span class="ml-4 text-success">
-        {ticket.price / 100} SEK
-      </span>
+      <Price price={ticket.price} class="ml-4" />
     </h1>
     <div
       class="flex flex-wrap items-center gap-2 text-sm text-base-content/60 *:flex *:items-center *:gap-1"
